+++ ricloud/samples/__init__.py
@@ -1,1 +1,2 @@
 from .icloud import icloud
+from .rirelay import rirelay