+++ ricloud/samples/icloud.py
@@ -4,22 +4,9 @@ from __future__ import absolute_import
 import click
 
 import ricloud
-from ricloud import conf
-from ricloud import storage
-from ricloud import utils
 
+from . import helpers
 from .utils import info, warn, success, prompt, pause, await_response
-
-
-def get_or_create_user(user_identifier):
-    info("Getting user...")
-
-    user_identifier = user_identifier or conf.get('samples', 'user_identifier')
-
-    user = ricloud.User.create(identifier=user_identifier)
-
-    success("User {} retrieved.".format(user.id))
-    return user
 
 
 def create_session(user_id, username, payload):
@@ -28,13 +15,10 @@ def create_session(user_id, username, payload):
     source_attributes = {
         "user": user_id,
         "type": "icloud.account",
-        "identifier": username
+        "identifier": username,
     }
 
-    session = ricloud.Session.create(
-        source=source_attributes,
-        payload=payload,
-    )
+    session = ricloud.Session.create(source=source_attributes, payload=payload)
 
     await_response(session)
 
@@ -43,15 +27,6 @@ def create_session(user_id, username, payload):
     else:
         success("Session {} created.".format(session.id))
 
-    return session
-
-
-def retrieve_session(session_id):
-    info("Retrieving session {}...".format(session_id))
-
-    session = ricloud.Session.retrieve(session_id)
-
-    success("Session {} retrieved.".format(session_id))
     return session
 
 
@@ -66,62 +41,6 @@ def handle_session_creation_failure(session, username, payload):
     raise click.Abort
 
 
-def await_poll(poll):
-    info("Awaiting poll completion...")
-
-    await_response(poll)
-
-    if poll.state == "failed":
-        warn("Poll failed with error: `{}`".format(poll.error))
-        raise click.Abort
-    else:
-        success("Poll {} completed.".format(poll.id))
-
-
-def create_poll(session, type, payload=None, source=None):
-    info("Creating account {} poll...".format(type))
-
-    request_payload = {"session": session, "type": type}
-
-    if payload:
-        request_payload["payload"] = payload
-
-    if source:
-        request_payload["source"] = source
-
-    poll = ricloud.Poll.create(**request_payload)
-
-    await_poll(poll)
-
-    return poll
-
-
-def process_poll_results(poll):
-    success("Processing results:")
-    results = {}
-    for result in poll.results:
-        identifier = result["identifier"]
-
-        success(" - identifier {}".format(identifier))
-        success("   stored at {}".format(result["url"]))
-
-        filename = utils.escape(identifier)
-        filename = utils.get_filename(poll.id, filename)
-
-        storage.download_result(result["url"], to_filename=filename)
-
-        if result["type"] == "json":
-            with open(filename, "rb") as f:
-                raw_result_data = f.read()
-            result_data = utils.decode_json(raw_result_data)
-            results[identifier] = result_data
-        else:
-            results[identifier] = filename
-
-        success("   downloaded locally to {}".format(filename))
-    return results
-
-
 def parse_file_ids_from_result_data(result_data):
     file_ids = []
     for data_entry in result_data["data"]:
@@ -134,31 +53,29 @@ def parse_file_ids_from_result_data(result_data):
 
 
 @click.command()
-@click.argument(
-    "icloud_username"
-)
+@click.argument("icloud_username")
 @click.password_option(
-    confirmation_prompt=False, help='The password of the iCloud account to access.'
+    confirmation_prompt=False, help="The password of the iCloud account to access."
 )
 @click.option(
     "--user_identifier",
     "user_identifier",
-    help='Optional. The `identifier` attribute of the User resource to associate this request with.'
+    help="Optional. The `identifier` attribute of the User resource to associate this request with.",
 )
 @click.option(
     "--session",
     "session_id",
-    help='Optional. The ID of the session to use for this request.'
+    help="Optional. The ID of the session to use for this request.",
 )
 def icloud(icloud_username, password, user_identifier=None, session_id=None):
-    """Sample implementation of the iCloud login and data retrieval process.
+    """Sample implementation for the iCloud service.
 
     Both ICLOUD_USERNAME and password are required.
     """
-    user = get_or_create_user(user_identifier)
+    user = helpers.get_or_create_user(user_identifier)
 
     if session_id:
-        session = retrieve_session(session_id)
+        session = helpers.retrieve_session(session_id)
     else:
         session_payload = {"password": password}
 
@@ -168,9 +85,9 @@ def icloud(icloud_username, password, user_identifier=None, session_id=None):
 
     account_data_payload = {"data_types": [data_type]}
 
-    poll = create_poll(session, "data", account_data_payload)
+    poll = helpers.create_poll(account_data_payload, session=session)
 
-    result_data = process_poll_results(poll)
+    result_data = helpers.process_poll_results(poll)
 
     file_ids = parse_file_ids_from_result_data(result_data[data_type])
 
@@ -179,8 +96,8 @@ def icloud(icloud_username, password, user_identifier=None, session_id=None):
 
         pause("Press any key to download the first 5 files...")
 
-        account_files_payload = {"file_ids": file_ids[:5]}
+        account_files_payload = {"files": file_ids[:5]}
 
-        poll = create_poll(session, "files", account_files_payload)
+        poll = helpers.create_poll(payload=account_files_payload, session=session)
 
-        process_poll_results(poll)
+        helpers.process_poll_results(poll)